feat(bridge-ui): Simplify the wallet connection modal
