+++ bridge-ui/package.json
@@ -22,10 +22,11 @@
   "dependencies": {
     "@consensys/linea-sdk": "0.3.0",
     "@headlessui/react": "2.1.9",
-    "@tanstack/react-query": "5.59.3",
+    "@tanstack/react-query": "5.62.16",
     "@wagmi/connectors": "5.1.15",
-    "@wagmi/core": "2.13.8",
-    "@web3modal/wagmi": "5.1.11",
+    "@wagmi/core": "2.16.3",
+    "@reown/appkit": "1.6.3",
+    "@reown/appkit-adapter-wagmi": "1.6.3",
     "clsx": "^2.1.1",
     "compare-versions": "6.1.1",
     "date-fns": "4.1.0",
@@ -44,8 +45,8 @@
     "sharp": "0.33.5",
     "swiper": "11.1.14",
     "tailwind-merge": "^2.5.3",
-    "viem": "2.21.19",
-    "wagmi": "2.12.17",
+    "viem": "2.22.4",
+    "wagmi": "2.14.6",
     "zustand": "4.5.4"
   },
   "devDependencies": {
